move node/zone/pool to separate header file
