+++ test/CMakeLists.txt
@@ -1,13 +1,17 @@
 # libs
 SET(TEST_NEEDED_LIBS ckit gtest curl ssl crypto z dl json11 boost_system boost_thread log4cplus)
 
-add_executable(test_util ${util_source} util/test_util.cpp)
+add_executable(test_util util/test_util.cpp)
 target_link_libraries(test_util ${TEST_NEEDED_LIBS} )
 add_test(test_util test_util)
 
-add_executable(test_consul_client ${all_source} balancer/test_consul_client.cpp)
+add_executable(test_consul_client balancer/test_consul_client.cpp)
 target_link_libraries(test_consul_client ${TEST_NEEDED_LIBS})
 add_test(test_consul_client test_consul_client)
+
+add_executable(test_consul_balancer balancer/test_consul_balancer.cpp)
+target_link_libraries(test_consul_balancer ${TEST_NEEDED_LIBS})
+add_test(test_consul_balancer test_consul_balancer)
 
 #if(UNITTEST)
 #    add_executable(test_util ${util_source} test/util/test_util.cpp)
